reogranised chess rust crate to be tidier
